Refactored code for optimization and code referencing.
