+++ src/main/java/net/crystudio/crystudiomod/Item/SickleItem.java
@@ -1,18 +1,20 @@
 package net.crystudio.crystudiomod.Item;
 
+import net.crystudio.crystudiomod.CustomHelpers.CustomItemHelper;
+import net.crystudio.crystudiomod.CustomToolMaterial.SickleToolMaterial;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
-import net.minecraft.block.Fertilizable;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.*;
 import net.minecraft.registry.tag.BlockTags;
-import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
+import net.minecraft.state.property.IntProperty;
+import net.minecraft.state.property.Properties;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
@@ -30,6 +32,7 @@ public class SickleItem extends MiningToolItem{
     private static final ToolMaterial material = new SickleToolMaterial();
     public static final Item Sickle = new SickleItem(material, material.getAttackDamage(), 0, new FabricItemSettings().maxCount(1));
     protected static final Map<Block, Pair<Predicate<ItemUsageContext>, Consumer<ItemUsageContext>>> HARVESTING_ACTIONS;
+    private static final IntProperty AGE = Properties.AGE_7;
 
     public SickleItem(ToolMaterial material, float attackDamage, float attackSpeed, Item.Settings settings) {
         super(attackDamage, attackSpeed, material, BlockTags.HOE_MINEABLE, new FabricItemSettings().maxCount(1));
@@ -38,18 +41,6 @@ public class SickleItem extends MiningToolItem{
     public static void Init () {
         CustomItemHelper.RegisterItem("sickle", Sickle);
         CustomItemHelper.AddToItemGroup(ItemGroups.TOOLS, Sickle);
-    }
-
-    private static void useOnFertilizable(ItemStack stack, World world, BlockPos pos) {
-        Fertilizable fertilizable;
-        BlockState blockState = world.getBlockState(pos);
-        if (blockState.getBlock() instanceof Fertilizable && (fertilizable = (Fertilizable)((Object)blockState.getBlock())).isFertilizable(world, pos, blockState, world.isClient)) {
-            if (world instanceof ServerWorld) {
-                if (fertilizable.canGrow(world, world.random, pos, blockState)) {
-                    fertilizable.grow((ServerWorld)world, world.random, pos, blockState);
-                }
-            }
-        }
     }
 
    @Override
@@ -81,12 +72,14 @@ public class SickleItem extends MiningToolItem{
        }
    }
 
-   public static Consumer<ItemUsageContext> dropItemProcedureWheat(BlockState result) {
+   public static Consumer<ItemUsageContext> dropItemProcedureWheat() {
        return (context) -> {
-           context.getWorld().setBlockState(context.getBlockPos(), result,Block.NOTIFY_ALL | Block.REDRAW_ON_MAIN_THREAD);
-           context.getWorld().emitGameEvent(GameEvent.BLOCK_CHANGE, context.getBlockPos(), GameEvent.Emitter.of(context.getPlayer(), result));
-           Block.dropStack(context.getWorld(), context.getBlockPos(), context.getSide(), new ItemStack(Items.WHEAT).copyWithCount(calculateWheatStackCountForFortune(context)));
-           useOnFertilizable(context.getStack(), context.getWorld(), context.getBlockPos());
+           int growthRange = (int)Math.round((Math.random() * 4) + 1);
+           World world = context.getWorld();
+           BlockState blockState = world.getBlockState(context.getBlockPos());
+           world.setBlockState(context.getBlockPos(), blockState.with(AGE, growthRange));
+           world.emitGameEvent(GameEvent.BLOCK_CHANGE, context.getBlockPos(), GameEvent.Emitter.of(context.getPlayer(), blockState));
+           Block.dropStack(world, context.getBlockPos(), context.getSide(), new ItemStack(Items.WHEAT).copyWithCount(calculateWheatStackCountForFortune(context)));
        };
    }
 
@@ -128,6 +121,6 @@ public class SickleItem extends MiningToolItem{
    }
 
    static {
-        HARVESTING_ACTIONS = Maps.newHashMap(ImmutableMap.of(Blocks.WHEAT, Pair.of(SickleItem::canHarvest, dropItemProcedureWheat(Blocks.WHEAT.getDefaultState())), Blocks.GRASS, Pair.of(SickleItem::canHarvest, dropItemProcedureGrass(Blocks.GRASS.getDefaultState())), Blocks.TALL_GRASS, Pair.of(SickleItem::canHarvest, dropItemProcedureGrass(Blocks.TALL_GRASS.getDefaultState()))));
+        HARVESTING_ACTIONS = Maps.newHashMap(ImmutableMap.of(Blocks.WHEAT, Pair.of(SickleItem::canHarvest, dropItemProcedureWheat()), Blocks.GRASS, Pair.of(SickleItem::canHarvest, dropItemProcedureGrass(Blocks.GRASS.getDefaultState())), Blocks.TALL_GRASS, Pair.of(SickleItem::canHarvest, dropItemProcedureGrass(Blocks.TALL_GRASS.getDefaultState()))));
    }
 }
